# headless/api/documentation/openapi.yaml
openapi: "3.0.0"
info:
  title: "phpBB 4 Headless API"
  version: "1.0.0"
  description: |
    RESTful API layer for phpBB 4.
    Provides all forum functionality via JSON for modern frontend applications.
  contact:
    name: "EBTURK"
servers:
  - url: "/api/v1"
    description: "API v1"
security:
  - BearerAuth: []
paths:
...
  /users/{user_id}:
    get:
      tags: [Users]
      summary: "User profile"
      operationId: "userShow"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "User profile"
    put:
      tags: [Users]
      summary: "Update user profile"
      operationId: "userUpdate"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Profile updated"
  /users/{user_id}/posts:
    get:
      tags: [Users]
      summary: "User posts"
      operationId: "userPosts"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Post list"
  /users/{user_id}/topics:
    get:
      tags: [Users]
      summary: "User topics"
      operationId: "userTopics"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Topic list"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT